backend/requirements.txt
fastapi[standard]
sqlalchemy
pydantic
passlib
pyjwt
pymysql